requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
plotly>=5.20
dash>=2.15
psycopg2-binary>=2.9
SQLAlchemy>=2.0
PyYAML>=6.0
Faker>=19.0
joblib>=1.3
matplotlib>=3.8
seaborn>=0.13
pytest>=7.4
python-dotenv>=1.0
pytz>=2024.1
